imagebitmap as default format. Support parse SVG to imagebitmap
