+++ modules/images/src/lib/parsers/parse-to-image-bitmap.js
@@ -1,19 +1,45 @@
 /* global createImageBitmap */
-import {getBlob} from './get-blob';
+import {isSVG, getBlob} from './svg-utils';
+import parseToImage from './parse-to-image';
 
-// Asynchronously parses an array buffer into an ImageBitmap - this contains the decoded data
-// Supported on worker threads, not supported on Edge, IE11 and Safari
-// https://developer.mozilla.org/en-US/docs/Web/API/ImageBitmap#Browser_compatibility
+const EMPTY_OBJECT = {};
 
 let imagebitmapOptionsSupported = true;
 
-// TODO - createImageBitmap supports source rect (5 param overload), pass through?
+/**
+ * Asynchronously parses an array buffer into an ImageBitmap - this contains the decoded data
+* ImageBitmaps are supported on worker threads, but not supported on Edge, IE11 and Safari
+* https://developer.mozilla.org/en-US/docs/Web/API/ImageBitmap#Browser_compatibility
+*
+* TODO - createImageBitmap supports source rect (5 param overload), pass through?
+*/
 export default async function parseToImageBitmap(arrayBuffer, options, url) {
-  const blob = getBlob(arrayBuffer, url);
+  let blob;
+
+  // Cannot parse SVG directly to ImageBitmap, parse to Image first
+  if (isSVG(url)) {
+    // Note: this only works on main thread
+    const image = await parseToImage(arrayBuffer, options, url);
+    blob = image;
+  }
+
+  // Create blob from the array buffer
+  if (!blob) {
+    blob = getBlob(arrayBuffer, url);
+  }
+
   let imagebitmapOptions = options && options.imagebitmap;
 
-  // Firefox crashes if imagebitmapOptions is supplied
-  // Avoid supplying if not provided, remember if not supported
+  return await safeCreateImageBitmap(blob, imagebitmapOptions);
+}
+
+/**
+ * Safely creates an imageBitmap with options
+ * *
+ * Firefox crashes if imagebitmapOptions is supplied
+ * Avoid supplying if not provided or supported, remember if not supported
+ */
+async function safeCreateImageBitmap(blob, imagebitmapOptions = null) {
   if (isEmptyObject(imagebitmapOptions) || !imagebitmapOptionsSupported) {
     imagebitmapOptions = null;
   }
@@ -30,8 +56,6 @@ export default async function parseToImageBitmap(arrayBuffer, options, url) {
 
   return await createImageBitmap(blob);
 }
-
-const EMPTY_OBJECT = {};
 
 function isEmptyObject(object) {
   for (const key in object || EMPTY_OBJECT) {
